AST: Move AvailabilityConstraint implementation to a separate file.
